feat: Promote the inline AST into Content (inline-ast Phase 2, step 1)
Move the Phase 1 tree-recorder machinery out of the test build into a
production module (`content::inline_tree`) that wraps the parser's own
inline renderer, and store the resulting inline node tree as a live
artifact on `Content` (`Content::inlines()`).

Tree building is gated behind an opt-in `Parser::with_inline_tree(true)`
flag. With it off (the default), the parse path is byte- and
performance-identical to before, so every existing `.rendered()`
assertion passes unchanged. With it on, each block's tree is built by a
counter-safe second substitution pass: `SubstitutionGroup::apply` clones
the parser before the authoritative pass advances any document counter,
runs the recording pipeline on the clone, and parses the recorded markers
into the tree — so footnotes, callouts, and counters are numbered
identically to the authoritative output.

The `inline_recorder` differential corpus now drives the production
module directly, so its byte-for-byte fold parity guards the shipped code.

This is the "promote the tree into Content" step of Phase 2; making
`rendered()` an authoritative fold and deleting the sentinel systems
remain (see the design doc update).
