+++ parser/src/content/mod.rs
@@ -10,6 +10,8 @@ pub(crate) use content::{
     rehome_xref_placeholders, render_xref_template, strip_footnote_marker_spans,
 };
 
+pub(crate) mod inline_tree;
+
 mod macros;
 pub(crate) use macros::apply_macros_with_leading_anchor_registered;
